+++ bot/src/main/java/edu/java/configuration/ApplicationConfig.java
@@ -3,7 +3,7 @@ package edu.java.configuration;
 import com.pengrad.telegrambot.TelegramBot;
 import com.pengrad.telegrambot.model.BotCommand;
 import com.pengrad.telegrambot.request.SetMyCommands;
-import edu.java.bot.commandParser.CommandParser;
+import edu.java.bot.command.CommandParser;
 import edu.java.bot.dict.CommandDict;
 import edu.java.bot.exception.BotExceptionHandler;
 import edu.java.bot.listener.BotUpdatesListener;
